jar: iterate over zip file headers instead of using fs.WalkDir
Avoid some of the edge cases we've hit when combining archive/zip and
fs.WalkDir and iterate over the ZIP files directly. Lets us drop some
workarounds.

Negative is that we have to break a public API, but hopefully that's
okay while we're pre v1.0.0.

PR attempts to make the diff as small as possible.