Allow JAR scanning to continue after error.
Adds a Parser.FileError handler similar to Walker.HandleError to allow
the caller to decide whether or not to abort the scan when processing
a file results in an error.

corrupt.jar and corrupt_jar_in_jar.jar contain vulnerable log4j
classes after a class with an unsupported compression type.  They will
only be detected as vulnerable if scanning continues past the error.