Remove unnecessary wildcards from pre-2.15 JndiManager constructor pattern
The previous pattern matching the pre-2.15 JndiManager constructor
uses wildcards to avoid three bytes that are problematic in regular
expressions.  binaryregexp does not have this limitation, so I've
replaced the three wildcards with their literal values.

I've tested this change on a corpus of all log4j 2 releases through
2.16 and verified that it does not change the classification decision
of any version.